// client/src/schema/userSchema.ts
import { z } from "zod";

export const userSignupSchema = z.object({
    fullname: z.string().min(1, "Fullname is required"),
    email: z.string().email("Invalid email"),
    password: z.string().min(6, "Password must be at least 6 characters"),
    contact: z.string().min(10, "Contact number must be 10 digit").max(10, "Contact number must be 10 digit")
})

export type SignupInputState = z.infer<typeof userSignupSchema>;


export const userLoginSchema = z.object({
    email: z.string().email("Invalid email"),
    password: z.string().nonempty({ message: "Password is required" }),
})

export type LoginInputState = z.infer<typeof userLoginSchema>;